feat: more fine-granular timer update (#94)
At present, it is up to the host app how often it calls
`update_timers_at`. Once a second should be enough but ideally, a `Tunn`
just exposes when the next internal timer fires. This is being tackled
more generally in #61. In order to ensure we send keep-alive messages in
time, we incorporate `want_passive_keepalive_at` into the
`next_timer_update` into it.