fix: return at most current time from `next_timer_update` (#104)
It is possible for `boringtun`'s timers to return a timestamp in the
past when the following happens:

- New session gets handshaked
- No traffic is sent for 2+ minutes
- Initiator sends a new packet

In this case, `boringtun` will want to re-key the session because 2
minutes have passed since the session got established. Most of
`boringtun`'s timers are dependent on the session start time, so in
several cases, it may happen that an inactive session suddenly returns a
timestamp in the past.

Usually, this isn't a problem as application SHOULD check for that and
treat it as a no-op (or immediate wake-up). However, I'd consider it to
be generally cleaner for the state machine to simply default to the last
known "now" time.