chore: log whether the expired session is the current one (#105)
WireGuard ensures perfect-forward secrecy by rolling over to a new
session every 2 minutes. To ensure this happens smoothly, an old session
isn't immediately discarded. This allows peers to still decrypt packets
from old sessions for an additional minute whilst the new session is
being established. Thus, packets already in-flight can still be handled.

An artifact of this designs is that the old session will then expire
with the timer and still print `SESSION_EXPIRED(REJECT_AFTER_TIME)`.
Seeing only this in the logs may be alerting when in fact it is
completely normal. To make this more obvious, we add a boolean
`is_current` to the log message that tells us whether the expired
session is the current one. If `is_current = false`, the log is
harmless.