chore: expose `preshared_key` from `NoiseParams` (#106)
In order to detect whether two tunnels are the exact same, we not only
need to compare the remote and our public key but also the
`preshared_key`. If it differs, the two nodes will not be able to make a
handshake despite the public keys being the same.

In order to check this, `Tunn` needs to expose the `preshared_key`.