feat: codify 24-bit connection index (#112)
In order to identify and correctly route packets for peers, the index we
generate for a `Tunn` instance is split into a 24-bit global index (the
higher 24 bits) and an 8-bit session index (the lower 8 bits). To make
this more obvious, we introduce an `Index` struct that codifies these
rules and also implements `fmt::Display` such that we display both
components of the index.